fix(canvas-workspace): extract reference-management hook to fix failing file-size gate
Workbench/index.tsx had grown to 516 lines, tripping the hard 500-line
governance test (baseline was 512). Per the documented frontend
convention ("lift state machines and side effects into a useXxxController
hook"), extract all reference-drawer/cross-workspace-reference state and
handlers into useReferenceManagement.ts, mirroring the existing
useChatInsertionBridge/useMountedWorkspaceIds pattern in the same folder.
Workbench/index.tsx is now 242 lines; the stale 512-line baseline entry
is removed since the file no longer needs the exception.
